Refactor ONNX support modules
